Updating the RCP example code and the RCP Beginners Guide.
A user tried to compile what was in the RCP Beginners Guide and it did not
compile.  I updated the example code and manually updated it in the appendix
to the RCP Beginners Guide.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework, Teuchos
0) MPI_DEBUG => passed: passed=139,notpassed=0 (0.81 min)
1) SERIAL_RELEASE => passed: passed=136,notpassed=0 (0.77 min)
Other local commits for this build/test group: b47a98c